Added DiagnosisCreater basic
Created another application in the same program called “ERSimAdmin”
which is designed to give a user interface to create diagnosis and
patient information is a somewhat user friendly way.

Currently the program displays the “trigger” list and based on which
trigger is selected displays the “responses” list on the right. NOTE:
the diagnosis is hard coded as id=1 currently but this should be
modify-able.